Move lock decorators to utils/lock.py
`LockGuard` was lonely and the decorators were cluttering up
decorators.py.
