+++ bot/decorators.py
@@ -1,26 +1,17 @@
 import asyncio
-import inspect
 import logging
 import typing as t
-from collections import defaultdict
 from contextlib import suppress
-from functools import partial, wraps
-from weakref import WeakValueDictionary
+from functools import wraps
 
 from discord import Member, NotFound
 from discord.ext.commands import Cog, Context, check
 
 from bot.constants import Channels, RedirectOutput
-from bot.errors import LockedResourceError
-from bot.utils import LockGuard, function
+from bot.utils import function
 from bot.utils.checks import in_whitelist_check, with_role_check, without_role_check
 
 log = logging.getLogger(__name__)
-__lock_dicts = defaultdict(WeakValueDictionary)
-
-_IdCallableReturn = t.Union[t.Hashable, t.Awaitable[t.Hashable]]
-_IdCallable = t.Callable[[function.BoundArgs], _IdCallableReturn]
-ResourceId = t.Union[t.Hashable, _IdCallable]
 
 
 def in_whitelist(
@@ -64,78 +55,6 @@ def without_role(*role_ids: int) -> t.Callable:
     async def predicate(ctx: Context) -> bool:
         return without_role_check(ctx, *role_ids)
     return check(predicate)
-
-
-def lock(namespace: t.Hashable, resource_id: ResourceId, *, raise_error: bool = False) -> t.Callable:
-    """
-    Turn the decorated coroutine function into a mutually exclusive operation on a `resource_id`.
-
-    If any other mutually exclusive function currently holds the lock for a resource, do not run the
-    decorated function and return None. If `raise_error` is True, raise `LockedResourceError` if
-    the lock cannot be acquired.
-
-    `namespace` is an identifier used to prevent collisions among resource IDs.
-
-    `resource_id` identifies a resource on which to perform a mutually exclusive operation.
-    It may also be a callable or awaitable which will return the resource ID given an ordered
-    mapping of the parameters' names to arguments' values.
-
-    If decorating a command, this decorator must go before (below) the `command` decorator.
-    """
-    def decorator(func: t.Callable) -> t.Callable:
-        name = func.__name__
-
-        @wraps(func)
-        async def wrapper(*args, **kwargs) -> t.Any:
-            log.trace(f"{name}: mutually exclusive decorator called")
-
-            if callable(resource_id):
-                log.trace(f"{name}: binding args to signature")
-                bound_args = function.get_bound_args(func, args, kwargs)
-
-                log.trace(f"{name}: calling the given callable to get the resource ID")
-                id_ = resource_id(bound_args)
-
-                if inspect.isawaitable(id_):
-                    log.trace(f"{name}: awaiting to get resource ID")
-                    id_ = await id_
-            else:
-                id_ = resource_id
-
-            log.trace(f"{name}: getting lock for resource {id_!r} under namespace {namespace!r}")
-
-            # Get the lock for the ID. Create a lock if one doesn't exist yet.
-            locks = __lock_dicts[namespace]
-            lock = locks.setdefault(id_, LockGuard())
-
-            if not lock.locked():
-                log.debug(f"{name}: resource {namespace!r}:{id_!r} is free; acquiring it...")
-                with lock:
-                    return await func(*args, **kwargs)
-            else:
-                log.info(f"{name}: aborted because resource {namespace!r}:{id_!r} is locked")
-                if raise_error:
-                    raise LockedResourceError(str(namespace), id_)
-
-        return wrapper
-    return decorator
-
-
-def lock_arg(
-    namespace: t.Hashable,
-    name_or_pos: function.Argument,
-    func: t.Callable[[t.Any], _IdCallableReturn] = None,
-    *,
-    raise_error: bool = False,
-) -> t.Callable:
-    """
-    Apply the `lock` decorator using the value of the arg at the given name/position as the ID.
-
-    `func` is an optional callable or awaitable which will return the ID given the argument value.
-    See `lock` docs for more information.
-    """
-    decorator_func = partial(lock, namespace, raise_error=raise_error)
-    return function.get_arg_value_wrapper(decorator_func, name_or_pos, func)
 
 
 def redirect_output(destination_channel: int, bypass_roles: t.Container[int] = None) -> t.Callable:
