+++ bot/utils/__init__.py
@@ -2,10 +2,9 @@ from abc import ABCMeta
 
 from discord.ext.commands import CogMeta
 
-from bot.utils.lock import LockGuard
 from bot.utils.redis_cache import RedisCache
 
-__all__ = ["CogABCMeta", "LockGuard", "RedisCache"]
+__all__ = ["CogABCMeta", "RedisCache"]
 
 
 class CogABCMeta(CogMeta, ABCMeta):
